The queue storedin the dtabase now stores the basic move object with the aircraft as the target motion object. This can be used to call the queue for the specific map and notifying all the command objects to tick the map, to allow for multi-tick commands

+++ ATC/src/main/java/dev/jamtech/ATC/Aircraft.java
@@ -4,7 +4,6 @@
  */
 package dev.jamtech.ATC;
 
-import java.util.HashSet;
 import lombok.Data;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.core.mapping.Document;
@@ -81,7 +80,7 @@ public class Aircraft extends MotionObjectAbstract {
         this.setHeight(12000);
         this.setvSpeed(0.0, 0.0);
         this.setAcceleration(0.0, 0.0);
-        this.setId(new ObjectId().getTimestamp()+Aircraft.getIdNum());
+        this.setId((int)new ObjectId().getTimestamp()+Aircraft.getIdNum());
     }
     
     public static int getIdNum()
